+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `transactionFeeCurrencyIsoCode` field to localPayment
 * Add `refundFromTransactionFeeAmount` field to localPayment
 * Add `refundFromTransactionFeeCurrencyIsoCode` field to localPayment
+* Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
 * Add `threeDSecureVersion`, `authenticationResponse`, `directoryResponse`, `cavvAlgorithm` and `dsTransactionId` to 3DS pass thru fields
 * Add `payerInfo` field to paymentMethodNonce
 * Add `roomTax` support on transaction sale