+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 * Add `PostalCodeIsRequiredForCardBrandAndProcessor` to validation errors
 * Add `PayPalHereDetails` to Transaction 
+* Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
+* Add `ThreeDSecureInfo` to `CreditCardVerification`
 
 ## 2.18.0
 * Add `captureId` field to localPayment
@@ -10,7 +12,6 @@
 * Add `transactionFeeCurrencyIsoCode` field to localPayment
 * Add `refundFromTransactionFeeAmount` field to localPayment
 * Add `refundFromTransactionFeeCurrencyIsoCode` field to localPayment
-* Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
 * Add `threeDSecureVersion`, `authenticationResponse`, `directoryResponse`, `cavvAlgorithm` and `dsTransactionId` to 3DS pass thru fields
 * Add `payerInfo` field to paymentMethodNonce
 * Add `roomTax` support on transaction sale