+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `PayPalHereDetails` to Transaction 
 * Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
 * Add `ThreeDSecureInfo` to `CreditCardVerification`
+* Add `GraphQLClient` to `BraintreeGateway` class
 
 ## 2.18.0
 * Add `captureId` field to localPayment