+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 2.19.0
 * Add `PostalCodeIsRequiredForCardBrandAndProcessor` to validation errors
 * Add `PayPalHereDetails` to Transaction 
 * Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`